Add Csv support (#17)
* Fix data type

* Add csv file support

* Pretty numbers

+++ ermitsognnede-functions/ermitsognnedeFunctions/ermitsognnedeFunctions/Models/DistrictDataJsonModel.cs
@@ -18,7 +18,7 @@ namespace ermitsognnedeFunctions.Models
         public int DistrictPopulationCount { get; set; }
 
         [JsonProperty("incidence")]
-        public int Incidence { get; set; }
+        public double Incidence { get; set; }
 
         [JsonProperty("new_infected_count")]
         public int NewInfectedCount { get; set; }
